requirements.txt
streamlit==1.41.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
pandas==2.2.2
sentence-transformers==3.3.1
python-docx==1.1.2
pypdf==5.1.0
faiss-cpu==1.8.0.post1


